# Fix #140 part 2 - detect NaN when intersection ray with AABB