--- a/src/ray/intersect_default.rs
+++ b/src/ray/intersect_default.rs
@@ -1,7 +1,11 @@
 //! This file contains the generic implementation of [`RayIntersection`]
 
 use super::Ray;
-use crate::{aabb::Aabb, bounding_hierarchy::BHValue, utils::fast_max};
+use crate::{
+    aabb::Aabb,
+    bounding_hierarchy::BHValue,
+    utils::{fast_max, has_nan},
+};
 
 /// The [`RayIntersection`] trait allows for generic implementation of ray intersection
 /// useful for our SIMD optimizations.
@@ -14,6 +18,14 @@ impl<T: BHValue, const D: usize> RayIntersection<T, D> for Ray<T, D> {
     fn ray_intersects_aabb(&self, aabb: &Aabb<T, D>) -> bool {
         let lbr = (aabb[0].coords - self.origin.coords).component_mul(&self.inv_direction);
         let rtr = (aabb[1].coords - self.origin.coords).component_mul(&self.inv_direction);
+
+        if has_nan(&lbr) | has_nan(&rtr) {
+            // Assumption: the ray is in the plane of an AABB face. Be consistent and
+            // consider this a non-intersection. This avoids making the result depend
+            // on which axis/axes have NaN (min/max in the code that follows are not
+            // commutative).
+            return false;
+        }
 
         let (inf, sup) = lbr.inf_sup(&rtr);
 
@@ -29,6 +41,10 @@ impl<T: BHValue, const D: usize> RayIntersection<T, D> for Ray<T, D> {
     default fn ray_intersects_aabb(&self, aabb: &Aabb<T, D>) -> bool {
         let lbr = (aabb[0].coords - self.origin.coords).component_mul(&self.inv_direction);
         let rtr = (aabb[1].coords - self.origin.coords).component_mul(&self.inv_direction);
+
+        if has_nan(&lbr) | has_nan(&rtr) {
+            return false;
+        }
 
         let (inf, sup) = lbr.inf_sup(&rtr);