I got a bit distracted over the past week but, when I took a look this afternoon, I found a complete solution for #140.

The change is to consider all rays that are in the plane of an AABB face as non-intersecting with the AABB, by detecting NaN. Previously, a subset of those rays were intersecting, typically with NaN/infinity/finite-but-wrong intersection slices. It seems very difficult to assign a correct intersection slice to those rays.

A NaN results when a coordinate of the ray origin matches the corresponding coordinate of an AABB face AND the ray's direction has a value of 0 for that axis. This causes 0/0 = NaN, which is detectable.

## Testing
- [x] The included test fails without this change
- [x] The included fuzzer change fails without this change

## Benchmarking
I made some effort to optimize the code. Here is a quick benchmark:
### No SIMD
```
cargo bench --features bench -- bench_intersect_12k_triangles_bvh
before: 363.77ns
after: 382.76ns
regression: 5.2%
```
### SIMD
```
cargo bench --features bench,simd -- bench_intersect_12k_triangles_bvh
before: 294.96ns
after: 304.78ns
regression: 3.3%
```

Fixes #140 (again)