@svenstaro 2025-01-27
Do you also want to add a CHANGELOG entry?